tasklet modifications for solid
