+++ java/src/main/java/com/codurance/training/tasks/TaskProcessor.java
@@ -1,12 +1,11 @@
 package com.codurance.training.tasks;
 
-import com.codurance.training.tasks.service.MenuService;
+import com.codurance.training.tasks.processor.MenuService;
 
 public class TaskProcessor {
 
     public static void main(String[] args) throws Exception {
         new MenuService().run();
     }
-
 
 }
